Split out passwordless actions
