+++ client/lib/signup/step-actions/index.js
@@ -674,34 +674,3 @@ export function isSiteTopicFulfilled( stepName, defaultDependencies, nextProps )
 		flows.excludeStep( stepName );
 	}
 }
-
-/**
- * Creates a user account and sends the user a verification code via email to confirm the account.
- * Returns the dependencies for the step.
- *
- * @param {Function} callback Callback function
- * @param {object}   data     POST data object
- */
-export function createPasswordlessUser( callback, { email } ) {
-	wpcom
-		.undocumented()
-		.usersEmailNew( { email }, null )
-		.then( response => callback( null, response ) )
-		.catch( err => callback( err ) );
-}
-
-/**
- * Verifies a passwordless user code.
- *
- * @param {Function} callback Callback function
- * @param {object}   data     POST data object
- */
-export function verifyPasswordlessUser( callback, { email, code } ) {
-	wpcom
-		.undocumented()
-		.usersEmailVerification( { email, code }, null )
-		.then( response =>
-			callback( null, { email, username: email, bearer_token: response.token.access_token } )
-		)
-		.catch( err => callback( err ) );
-}
